added bulk peer

+++ Src/Services/Api/Wireguard.Api/Data/Common/EntityBase.cs
@@ -7,5 +7,5 @@ public class EntityBase
     public DateTime? CreateDate { get; set; } = DateTime.UtcNow;
     public DateTime? UpdateDate { get; set; } = DateTime.UtcNow;
 
-    [Key] public long Id { get; set; }
+    [Key] public int Id { get; set; }
 }
